Refactor using stock types
