+++ frontend/src/contexts/StockTransactionsContext.tsx
@@ -6,19 +6,7 @@ import React, {
   ReactNode,
 } from 'react'
 import { useAuth } from './AuthContext'
-
-export interface StockTransactions {
-  stock_tx_id?: string
-  wallet_tx_id?: string
-  stock_id?: number
-  order_status?: string
-  is_buy?: boolean
-  order_type?: string
-  stock_price?: number
-  quantity?: number
-  time_stamp?: string
-  __v?: number
-}
+import { StockTransactions } from '../types/stocks'
 
 type StockTransactionsType = {
   stock_transactions: StockTransactions[]
